# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Changer REM16 Framework API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SurveyInputThinkingFocus =
  (typeof SurveyInputThinkingFocus)[keyof typeof SurveyInputThinkingFocus];

export const SurveyInputThinkingFocus = {
  Proof: "Proof",
  Process: "Process",
  People: "People",
  Possibility: "Possibility",
} as const;

export type SurveyInputOrientation =
  (typeof SurveyInputOrientation)[keyof typeof SurveyInputOrientation];

export const SurveyInputOrientation = {
  Eager: "Eager",
  Cautious: "Cautious",
} as const;

export type SurveyInputChangeRole =
  (typeof SurveyInputChangeRole)[keyof typeof SurveyInputChangeRole];

export const SurveyInputChangeRole = {
  Rockstar: "Rockstar",
  Roadie: "Roadie",
} as const;

export interface SurveyInput {
  stakeholderName: string;
  role: string;
  thinkingFocus: SurveyInputThinkingFocus;
  orientation: SurveyInputOrientation;
  changeRole: SurveyInputChangeRole;
  projectId?: number | null;
}

export interface SurveyResponse {
  id: number;
  stakeholderName: string;
  role: string;
  thinkingFocus: string;
  orientation: string;
  changeRole: string;
  mentalModel: string;
  mentalModelDescription: string;
  projectId?: number | null;
  createdAt: string;
}

export interface ProjectInput {
  name: string;
  bcipCanvas?: string | null;
  changeLogic?: string | null;
  changeStrategy?: string | null;
  managerName?: string | null;
}

export interface ProjectResponse {
  id: number;
  name: string;
  bcipCanvas?: string | null;
  changeLogic?: string | null;
  changeStrategy?: string | null;
  managerName?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface GenerateMessageInput {
  surveyId: number;
  projectId: number;
}

export type UpdateMessageInputStatus =
  (typeof UpdateMessageInputStatus)[keyof typeof UpdateMessageInputStatus];

export const UpdateMessageInputStatus = {
  draft: "draft",
  approved: "approved",
  sent: "sent",
} as const;

export interface UpdateMessageInput {
  editedContent?: string | null;
  status?: UpdateMessageInputStatus;
}

export interface MessageResponse {
  id: number;
  surveyId: number;
  projectId: number;
  stakeholderName: string;
  mentalModel: string;
  generatedContent: string;
  editedContent?: string | null;
  status: string;
  createdAt: string;
  updatedAt: string;
}

export interface MentalModelCount {
  mentalModel: string;
  count: number;
}

export interface DashboardStats {
  totalSurveys: number;
  totalProjects: number;
  totalMessages: number;
  approvedMessages: number;
  mentalModelDistribution: MentalModelCount[];
}

export type GetSurveys200 = {
  surveys: SurveyResponse[];
};

export type GetProjects200 = {
  projects: ProjectResponse[];
};

export type GetMessages200 = {
  messages: MessageResponse[];
};