# lib/api-zod/src/generated/types/mentalModelCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Changer REM16 Framework API
 * OpenAPI spec version: 0.1.0
 */

export interface MentalModelCount {
  mentalModel: string;
  count: number;
}