# lib/api-zod/src/generated/types/projectResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Changer REM16 Framework API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectResponse {
  id: number;
  name: string;
  bcipCanvas?: string | null;
  changeLogic?: string | null;
  changeStrategy?: string | null;
  managerName?: string | null;
  createdAt: Date;
  updatedAt: Date;
}