# lib/api-zod/src/generated/types/surveyInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Changer REM16 Framework API
 * OpenAPI spec version: 0.1.0
 */
import type { SurveyInputChangeRole } from "./surveyInputChangeRole";
import type { SurveyInputOrientation } from "./surveyInputOrientation";
import type { SurveyInputThinkingFocus } from "./surveyInputThinkingFocus";

export interface SurveyInput {
  stakeholderName: string;
  role: string;
  thinkingFocus: SurveyInputThinkingFocus;
  orientation: SurveyInputOrientation;
  changeRole: SurveyInputChangeRole;
  projectId?: number | null;
}