# lib/api-zod/src/generated/types/surveyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Game Changer REM16 Framework API
 * OpenAPI spec version: 0.1.0
 */

export interface SurveyResponse {
  id: number;
  stakeholderName: string;
  role: string;
  thinkingFocus: string;
  orientation: string;
  changeRole: string;
  mentalModel: string;
  mentalModelDescription: string;
  projectId?: number | null;
  createdAt: Date;
}